requirements.txt
earthengine-api==0.1.396
folium==0.17.0
google-generativeai==0.7.0
python-dotenv==1.0.1